+++ Assets/Scripts/FirstPersonController.cs
@@ -415,7 +415,7 @@ public class FirstPersonController : MonoBehaviour
             ball.TryKick(gameObject);
     }
 
-    private const string GameVersion = "0.3.204-dev";
+    private const string GameVersion = "0.3.205-dev";
 
     private float lastSpeed;
     private bool lastSprinting;